Connection String Modified for Order Service (#184)
* Link missing for Deploy to Azure

If I push the link of the "Deploy to Azure" button, it failed to show the Form in the Azure Portal.

And it seems that the link had changed as follows.

Previous Link (Currently Failed)
https://raw.githubusercontent.com/Azure-Samples/acme-fitness-store/Azure/workshops/azure-spring-apps-enterprise/full/03-workshop-environment-setup/acmedeploy.json)

Now (Correct Link)
https://raw.githubusercontent.com/Azure-Samples/acme-fitness-store/Azure/azure-spring-apps-enterprise/workshops/azure-spring-apps-enterprise/full/03-workshop-environment-setup/acmedeploy.json

So I modified the link for it.

* Connection String Modified for Order Service

If I configured the Connection String for Npgsql (C#).
Following error was showed and couldn't start the order service.

So I modified to be able to access to the PostgreSQL from the catalog service.

```bash
az keyvault secret set \
    --vault-name acme-fitness-kv-yoshio \
    --name "ConnectionStrings--OrderContext" \
    --value "Server=$POSTGRES_SERVER.postgres.database.azure.com;Database=$ORDER_SERVICE_DB;Port=5432;SSL Mode=Require;Trust Server Certificate=true;User ID=$POSTGRES_SERVER_USER;Password=$POSTGRES_SERVER_PASSWORD"
```

```text
Unhandled exception. System.ArgumentException: To validate server certificates, please use VerifyFull or VerifyCA instead of Require. To disable validation, explicitly set 'Trust Server Certificate' to true. See https://www.npgsql.org/doc/release-notes/6.0.html for more details.
```

* Modified Env Value

I modified the expression of the env variables

* I modified the style with more visible

I modified the style with more visible.

I also tried to the separate line and works fine.